Refactor code for common internal api’s and make separate files for ChainlinkInitServer and ChainlinkServer. added support for custom username and api for chainlink node
