+++ seedemu/services/ChainlinkService/ChainlinkTemplates/ChainlinkFileTemplate.py
@@ -1,0 +1,36 @@
+from typing import Dict
+
+# Templates for configuration files remain the same
+ChainlinkFileTemplate: Dict[str, str] = {}
+
+ChainlinkFileTemplate['config'] = """\
+[Log]
+Level = 'info'
+
+[WebServer]
+AllowOrigins = '*'
+SecureCookies = false
+
+[WebServer.TLS]
+HTTPSPort = 0
+
+[[EVM]]
+ChainID = '1337'
+
+[[EVM.Nodes]]
+Name = 'SEED Emulator'
+WSURL = 'ws://{ip_address}:8546'
+HTTPURL = 'http://{ip_address}:8545'
+"""
+
+ChainlinkFileTemplate['secrets'] = """\
+[Password]
+Keystore = 'mysecretpassword'
+[Database]
+URL = 'postgresql://postgres:mysecretpassword@localhost:5432/postgres?sslmode=disable'
+"""
+
+ChainlinkFileTemplate['api'] = """\
+{username}
+{password}
+"""
