--- seedemu/services/ChainlinkService/ContractTemplates/__init__.py
@@ -1,2 +1,0 @@
-from .LinkTokenDeploymentTemplate import LinkTokenDeploymentTemplate
-from .OracleContractDeploymentTemplate import OracleContractDeploymentTemplate
