+++ seedemu/services/ChainlinkService/__init__.py
@@ -2,5 +2,7 @@
 # encoding: utf-8
 # __author__ = 'Demon'
 from __future__ import annotations
-from .ContractTemplates import LinkTokenDeploymentTemplate, OracleContractDeploymentTemplate
-
+from .ChainlinkServerCommands import *
+from .ChainlinkInitializerServer import *
+from .ChainlinkServer import *
+from .ChainlinkTemplates import *
